WIP Example of getting class data by module
This adds methods to the jruby-puppet-service that allow for getting
environment class data grouped by module instead of all together in one list.
The intent would be to enable us to associate classes with modules, since
puppetdb doesn't store data about what module a class comes from. This would
give us some means of looking it up, though we'd probably want to reformat the
data more than this for that purpose; this is just a demonstration that it's
possible to get it at all.